Fixed redundant cast to float warning

+++ trunk/classes/lejos/robotics/proposal/InertialMoveProvider.java
@@ -137,7 +137,7 @@ public class InertialMoveProvider implements MoveProvider1
      */
     public float getAngularVelocity()
     {
-        return (float)gyro.getAngularVelocity();
+        return gyro.getAngularVelocity();
     }
 
     /**
